pyproject.toml
[tool.poetry]
name = "auction-app"
version = "0.1.0"
description = ""
authors = ["MKurziankou <MKurzenkov.otd14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
load-dotenv = "^0.1.0"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.3.6"
redis = "^5.0.1"
flower = "^2.0.1"
channels = "^4.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
